Remove dados chumbados e adiciona modulos reais de backend
Substitui os pontos ainda cobertos por dados fictícios no frontend por
integrações de verdade, adicionando onde faltava suporte no backend:
FichaTecnicaItem (receita/custo real de produção), PedidoCompra, Entrega/
Entregador (delivery), Notificacao (eventos reais de venda, estoque,
pagamento e pedido de compra), Despesa, Dashboard (agregação real de
vendas/estoque) e IntegracaoConfig (credenciais de integrações externas).
Também liga o acúmulo real de cashback em VendaService e libera o
autoatendimento da própria empresa em /empresa/minha-empresa.

+++ src/main/java/com/easy/eats/cashback/model/CashbackConfig.java
@@ -1,5 +1,7 @@
 package com.easy.eats.cashback.model;
 
+import java.math.BigDecimal;
+
 import com.easy.eats.empresa.model.model.Empresa;
 
 import jakarta.persistence.Entity;
@@ -31,9 +33,9 @@ public class CashbackConfig {
 
     @NotNull(message = "O percentual de acúmulo é obrigatório")
     @PositiveOrZero(message = "O percentual não pode ser negativo")
-    private Double percentualAcumulo;
+    private BigDecimal percentualAcumulo;
 
-    private Double valorMinimoParaAcumular;
+    private BigDecimal valorMinimoParaAcumular;
     private Boolean flAtivo;
 
     @ManyToOne
