+++ src/main/java/com/easy/eats/venda/quicksort/ProdutoRanking.java
@@ -1,5 +1,7 @@
 package com.easy.eats.venda.quicksort;
 
+import java.math.BigDecimal;
+
 import lombok.AllArgsConstructor;
 import lombok.Getter;
 import lombok.NoArgsConstructor;
@@ -13,6 +15,6 @@ public class ProdutoRanking {
 
     private String nomeProduto;
     private Double quantidadeVendida;
-    private Double faturamentoTotal;
+    private BigDecimal faturamentoTotal;
 
 }
